PVS/V1001: Add cast to return statement

+++ src/nvim/api/buffer.c
@@ -1535,7 +1535,7 @@ Integer nvim_buf_add_highlight(Buffer buffer,
                       (int)line, (colnr_T)col_start,
                       end_line, (colnr_T)col_end,
                       decoration_hl(hl_id), kExtmarkUndo);
-  return ns_id;
+  return (Integer)ns_id;
 }
 
 /// Clears namespaced objects (highlights, extmarks, virtual text) from
